Correct typos.
Only one or two true typos.

I left "licence" as it's technically valid - just not correct in North America where my bias lies.
However, "organize" wins over "organise" as the former is the more historically correct term across all variants of English, with the latter only coming in "recently"...  If you want, feel free to revert this change - not going to hurt my pinky fingers!
I could not find a widespread preference for "decentralised" so I left it as is - despite my bias.